move ApiContext from shared to features to later improve types ('metier')
