+++ webapp/src/features/api/client.ts
@@ -1,0 +1,47 @@
+import { API_URL } from "@shared/api/client";
+
+const fetchWithAuth = async (
+  endpoint: string,
+  options: RequestInit = {},
+  authToken: string | null,
+) => {
+  const headers = new Headers(options.headers);
+  headers.set("Authorization", authToken ? `Bearer ${authToken}` : "");
+
+  const res = await fetch(`${API_URL}${endpoint}`, {
+    ...options,
+    headers,
+  });
+
+  if (!res.ok) {
+    console.error(`Erreur API: ${res.status} ${res.statusText}`);
+    const errorData = await res.json().catch(() => ({
+      details: [res.statusText],
+      error: "Erreur de communication",
+    }));
+    console.error("Détails de l'erreur:", JSON.stringify(errorData, null, 2));
+
+    const error = new Error(`Erreur API: ${res.status}`);
+
+    // @ts-expect-error Property 'response' does not exist on type 'Error'.
+    error.response = { data: errorData };
+
+    throw error;
+  }
+
+  return res;
+};
+
+const fetchJSONWithAuth = async (
+  endpoint: string,
+  options: RequestInit = {},
+  authToken: string | null,
+) => (await fetchWithAuth(endpoint, options, authToken)).json();
+
+export const createApiClient = (authToken: string | null) => ({
+  // Bases
+  fetchDashboardData: (layerId: string) =>
+    fetchJSONWithAuth(`/maps/dashboard/${layerId}`, {}, authToken),
+});
+
+export type ApiClient = ReturnType<typeof createApiClient>;
